refactor: extract routes governor cut one
