+++ server/condition-welcome.ts
@@ -17,16 +17,6 @@
  * Pure — no DB, no model. Unit-tested.
  */
 
-const MEDICATION_SIGNAL = /\b(on medication|taking medication|my medication|my meds|my pills|blood thinners|antiretroviral|ARVs?|antiretrovirals?|insulin|metformin|warfarin|blood pressure (pills?|medication|tablets?)|epilepsy (medication|tablets?|pills?)|seizure medication|newly diagnosed|just diagnosed|just found out i have|blood test results?|doctor said i have|specialist said)\b/i;
-
-const CHRONIC_CONDITION_SIGNAL = /\b(i have diabetes|i.?m diabetic|pre.?diabetic|type [12] diabetes|my blood sugar|i have hypertension|i.?m hypertensive|my blood pressure is|i have (heart disease|a heart condition|kidney disease|liver disease|thyroid|pcos|epilepsy|hiv|aids))\b/i;
-
-/** Did they just disclose a condition or medication? */
-export function mentionsConditionOrMedication(message: string): boolean {
-  const m = message || "";
-  return MEDICATION_SIGNAL.test(m) || CHRONIC_CONDITION_SIGNAL.test(m);
-}
-
 /**
  * The reply: welcome first, boundary second, back to coaching third. It never names their
  * condition back at them, never comments on medicine, and never claims anything is "safe for"
