+++ server/handlers/safety.ts
@@ -22,6 +22,15 @@ import { asksForExport, formatExport } from "../data-export";
 import { sastDayKey } from "../sast";
 import { logChat } from "./chat-log";
 
+const MEDICATION_SIGNAL = /\b(on medication|taking medication|my medication|my meds|my pills|blood thinners|antiretroviral|ARVs?|antiretrovirals?|insulin|metformin|warfarin|blood pressure (pills?|medication|tablets?)|epilepsy (medication|tablets?|pills?)|seizure medication|newly diagnosed|just diagnosed|just found out i have|blood test results?|doctor said i have|specialist said)\b/i;
+const CHRONIC_CONDITION_SIGNAL = /\b(i have diabetes|i.?m diabetic|pre.?diabetic|type [12] diabetes|my blood sugar|i have hypertension|i.?m hypertensive|my blood pressure is|i have (heart disease|a heart condition|kidney disease|liver disease|thyroid|pcos|epilepsy|hiv|aids))\b/i;
+
+/** Safety owns recognition of medical disclosures; the welcome copy remains separate. */
+export function mentionsConditionOrMedication(message: string): boolean {
+  const m = message || "";
+  return MEDICATION_SIGNAL.test(m) || CHRONIC_CONDITION_SIGNAL.test(m);
+}
+
 // Send a Twilio message with exponential-backoff retries. On complete failure,
 // records a row in adminEvents so the coach can find missed alerts on reload.
 async function sendAlertWithRetry(opts: {
