+++ server/handlers/steps.ts
@@ -1,11 +1,17 @@
 import { db } from "../db";
 import { turnMutation } from "./chat-log";
-import { stepLogs } from "../../shared/schema";
+import { users, stepLogs } from "../../shared/schema";
 import { eq, desc, and, gte, lt } from "drizzle-orm";
 import { neverSilentLine } from "../reply-hygiene";
 import { educationNote } from "../education";
 import { stepBurnKcal } from "../targets";
-import { sastDayStart } from "../utils";
+import { isFutureIntent, isRetroactiveMeal, mealDateLabel, mentionsNotDone, parseMealDate, reportedInSomeClause, sastDayStart } from "../utils";
+import { checkPerfectDay } from "./checks";
+import { detectStepLog } from "../understanding/messy-intake";
+import { getDailyStepContext } from "../targets";
+import { getTodayWorkoutState } from "../workout-state";
+import { invalidatePatternCache } from "../cache";
+import { logChat, turnAlreadyWrote } from "./chat-log";
 
 /**
  * THE STEP WRITE — all of it, for every door (2026-07-19; bypass removed 2026-08-26, issue #63).
@@ -145,4 +151,119 @@ export function getStepResponse(steps: number, target: number, weightKg = 75, st
   // them and the signature is not the point; the VOICE was the point, and it has one owner.
   void target; void weightKg; void streak; void weeklyAvg; void user; void isWorkoutDay;
   return neverSilentLine("steps", { amount: steps.toLocaleString("en-ZA") });
+}
+
+
+/** The route adapter for the step writer: it preserves turn composition callbacks. */
+export async function handleStepReport(ctx: {
+  message: string; m: string; user: any; normalizedQuestion: boolean;
+  commitStep: (reply: string) => void; mayEndTurn: (owner: string) => boolean;
+  closeCoachingTurn: (reply: string) => string | Promise<string>; hasStepPart: () => boolean;
+}): Promise<string | null> {
+  const { message, m, user, normalizedQuestion } = ctx;  // ---- STEP LOG DETECTION (direct — no GPT cost) ----
+  // NOTE: If message also contains food (e.g. voice note: "I had eggs for breakfast and walked 3000 steps"),
+  // we log steps but do NOT return early — let it fall through to food scanning
+  // "12k steps", "8.5k steps", "12,000 steps", "12000 steps" — all valid
+  // Also: "Fitbit says 8500", "health app: 9000", "steps today: 7500"
+  // ALL STEP PARSING HAS ONE OWNER (Cut 5b). ~50 lines of regexes and the number arithmetic
+  // lived here, beside messy-intake's own extractStepCount — two step parsers for one fact, in
+  // two files. Moved whole; the guards that need this function's context stay here.
+  let sd = detectStepLog(m);
+  // A QUESTION IN ONE CLAUSE DOES NOT ERASE A REPORT IN ANOTHER (2026-08-26, issue #63). On
+  // "walked 8000 steps. what should I eat?" the extractor found 8 000 and threw it away, because
+  // isQuestionForm matches "should i" ANYWHERE in the bubble. Re-read per clause, only when the
+  // whole-message read already said no, and the clause's own parse becomes the parse — so the
+  // number stays tied to the sentence that reported it. See tracking-contract-tests, LAW 5.
+  if (!sd.loggableByForm) {
+    const stepClause = reportedInSomeClause(m, c => {
+      const d = detectStepLog(c);
+      return d.matched && d.loggableByForm && d.isExplicitLog;
+    });
+    if (stepClause) sd = detectStepLog(stepClause);
+  }
+  // Future-intent guard: "I'll walk 10k tomorrow" slips past the question check — must not log
+  // today. Explicit "walked 8,000 steps" is a log even if the classifier tagged the note a
+  // QUESTION because they also said "I'm exhausted" (live 2026-08-19 mixed note — steps dropped).
+  if (sd.loggableByForm && !turnAlreadyWrote("steps") && !isFutureIntent(m) && !mentionsNotDone(m) && (sd.isExplicitLog || !normalizedQuestion) && sd.matched) {
+    let steps = sd.steps;                       // a "8000 not 5000" correction rewrites it below
+    const stepHasMovementSignal = sd.hasMovementSignal;
+    if (!isNaN(steps) && steps > 0 && steps <= 100 && stepHasMovementSignal) {
+      return `That step count looks low — did the message cut off? Send your actual count, e.g. "8500 steps" or "walked 5km".`;
+    }
+    if (!isNaN(steps) && steps > 100 && steps < 100000) {
+      // Weekly AVERAGE reports ("my average this week is 6,400") are a summary, not
+      // today's count — logging them as today corrupts the day AND the 7-day trend.
+      // Coach on the week instead; clients may opt to report a weekly average only.
+      if (/\b(average|avg)\b/i.test(m) || (/\b(this|last|past)\s+week(?:ly)?\b/i.test(m) && !/\btoday\b/i.test(m))) {
+        const wkTarget = user.stepsTarget || 8500;
+        const wkDiff = steps - wkTarget;
+        const wkReply = `Weekly average noted: *${steps.toLocaleString()} steps/day* vs your ${wkTarget.toLocaleString()} target — ${wkDiff >= 0 ? "on target. Strong week 🔥" : `${Math.abs(wkDiff).toLocaleString()} short. One 15-minute walk a day closes that.`}\n\n_Daily counts or a weekly-average screenshot both work — whichever is easier for you._`;
+        await logChat(user.id, message, wkReply, "STEP_WEEKLY_REPORT");
+        return wkReply;
+      }
+      const baseStepsTarget = user.stepsTarget || 8500;
+      // Detect whether client already worked out today so we can ease step demand.
+      let workedOutToday = false;
+      try { workedOutToday = (await getTodayWorkoutState(user)).alreadyDoneToday; } catch { /* non-critical */ }
+      const { target, goalContext: stepGoalCtx } = getDailyStepContext(
+        baseStepsTarget, user.goalType || "fat_loss", workedOutToday
+      );
+      const stepIsRetro = isRetroactiveMeal(message);
+      const stepLoggedAt = stepIsRetro ? parseMealDate(message) : new Date();
+      // Allow a downward CORRECTION ("8000 steps not 50000", "wrong, 6k steps") to overwrite the
+      // day's count. Normally we keep only the HIGHER number (clients re-log a growing daily
+      // total), but an explicit correction must win in either direction. For "X not Y", X is the
+      // affirmed value — pull it out so the position of "steps" in the sentence doesn't matter.
+      const stepNotMatch = m.match(/\b([\d,]+)(\s*k)?\s*(?:steps?|staps?)?\s+not\s+[\d,]+/i);
+      if (stepNotMatch) {
+        const corrected = Math.round(parseFloat(stepNotMatch[1].replace(/,/g, "")) * (stepNotMatch[2] ? 1000 : 1));
+        if (corrected > 100 && corrected < 100000) steps = corrected;
+      }
+      const isStepCorrection = !!stepNotMatch
+        || /\b(wrong|actually|correction|i\s+meant|meant|should\s+be|mistake|typo|miscount|oops|my\s+bad)\b/i.test(m);
+      // ONE WRITE OWNER FOR THE STEP FACT (2026-08-26, issue #63). This built its own day-window
+      // query and upsert — twenty lines that logStepsForUser already was, and steps.ts said so at
+      // the top of itself: "mirroring the routes.ts inline upsert exactly ... Additive, routes
+      // keeps its own path." A mirror is a copy, and this one had already drifted on the half that
+      // reaches the client: the owner returns the count now STORED, while this path carried the
+      // count from the MESSAGE into the reply. With 9 000 already logged, "walked 3000 steps
+      // today" correctly left the row at 9 000 and answered "3 000 steps — nice one", quoting a
+      // number the ledger does not hold. So the write moves to the owner AND the answer is built
+      // from what the owner returns — that second half is the only reason the bypass mattered.
+      const storedSteps = await logStepsForUser(user.id, steps, { correction: isStepCorrection, at: stepLoggedAt });
+      await db.update(users).set({ lastActiveAt: new Date() }).where(eq(users.phoneNumber, phone));
+      invalidatePatternCache(user.id);
+      const sevenDaysAgo = new Date(Date.now() - 7 * 86_400_000);
+      const [perfectDay, streak, recentStepLogs] = await Promise.all([
+        checkPerfectDay(user.id, user.proteinTarget || 120, target),
+        getStepStreak(user.id),
+        db.select({ steps: stepLogs.steps }).from(stepLogs)
+          .where(and(eq(stepLogs.userId, user.id), gte(stepLogs.loggedAt, sevenDaysAgo)))
+          .orderBy(desc(stepLogs.loggedAt))
+          .limit(7),
+      ]);
+      // Divide by days actually logged, not a flat 7 — a new client who logged 9k steps
+      // on each of their first 3 days was told "7-day average: 3,857, below target".
+      const weeklyAvg = recentStepLogs.length >= 3
+        ? Math.round(recentStepLogs.reduce((s, r) => s + r.steps, 0) / recentStepLogs.length)
+        : undefined;
+      void stepGoalCtx; // used by getStepResponse via user.goalType
+      const stepReply = getStepResponse(storedSteps, target, parseFloat(user.currentWeight as string || "75") || 75, streak, weeklyAvg, user, workedOutToday);
+      const stepRetroNote = stepIsRetro ? `\n_Logged to ${mealDateLabel(stepLoggedAt)}._` : "";
+      const stepPart = (isStepCorrection ? `Fixed ✅ — step count updated to *${storedSteps.toLocaleString()}*.\n\n` : "") + stepReply + stepRetroNote + (perfectDay || "");
+      ctx.commitStep(stepPart);
+      await logChat(user.id, message, stepPart, "STEP_LOG");
+
+      // ONE RULE, NOT A PAIR BRANCH: this asked `alsoHasFood` with its own 30-word food regex,
+      // a second list of which facts were allowed to coexist with steps. The ledger knows.
+      if (ctx.mayEndTurn("steps")) return ctx.closeCoachingTurn(stepPart);
+    }
+  }
+
+  // Voice cut off: "I've walked..." with no number. Do not drop the walk.
+  if (!ctx.hasStepPart() && mentionedWalkWithoutCount(message) && !(m.includes("?") || sd.isQuestionForm) && !isFutureIntent(m)) {
+    ctx.commitStep(`Heard you walked — send the step count (e.g. "3000 steps") and I'll log it.`);
+  }
+
+  return null;
 }
